+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-data-export
 
 ## 1.1.0 (IN PROGRESS)
+* Populate JobId field on Job logs, update `stripes-smart-components` to `v3.0.0` to avoid errors. UIDEXP-35. 
 
 ## [1.0.0](https://github.com/folio-org/ui-data-export/tree/v1.0.0) (2020-03-13)
 * UI app created. Refs UIDEXP-8.